💄 Correct provider display names. Clean up provider details
