+++ components/pages/login/index.tsx
@@ -2,14 +2,7 @@ import { css } from '@emotion/core';
 import urlJoin from 'url-join';
 
 import PageLayout from '../../PageLayout';
-import {
-  GoogleLogo,
-  FacebookLogo,
-  GitHubLogo,
-  OrcidLogo,
-  LinkedInLogo,
-  Illustration,
-} from '../../theme/icons';
+import { Illustration } from '../../theme/icons';
 
 import { IconProps } from '../../theme/icons/types';
 import { getConfig } from '../../../global/config';
@@ -17,6 +10,8 @@ import { getConfig } from '../../../global/config';
 import { usePageQuery } from '../../../global/hooks/usePageContext';
 import { trim } from 'lodash';
 import ErrorNotification from '../../ErrorNotification';
+import providerMap, { ProviderDetail } from '../../../global/utils/providerTypeMap';
+import { ProviderType } from '../../../global/types';
 
 const LoginButton = ({
   Icon,
@@ -88,30 +83,6 @@ const LoginButton = ({
       </div>
     </a>
   );
-};
-
-enum ProviderType {
-  GOOGLE = 'GOOGLE',
-  ORCID = 'ORCID',
-  LINKEDIN = 'LINKEDIN',
-  GITHUB = 'GITHUB',
-  // FACEBOOK = 'FACEBOOK'
-}
-
-type ProviderDetail = {
-  displayName: string;
-  path: string;
-  icon: any;
-};
-type ProviderMap = { [k in ProviderType]: ProviderDetail };
-
-const providerMap: ProviderMap = {
-  [ProviderType.GOOGLE]: { displayName: 'Google', path: 'google', icon: GoogleLogo },
-  [ProviderType.ORCID]: { displayName: 'ORCiD', path: 'orcid', icon: OrcidLogo },
-  [ProviderType.GITHUB]: { displayName: 'GitHub', path: 'github', icon: GitHubLogo },
-  [ProviderType.LINKEDIN]: { displayName: 'LinkedIn', path: 'linkedin', icon: LinkedInLogo },
-  // Facebook will be hidden until provider implementation is fixed in Ego https://github.com/overture-stack/ego/issues/555
-  // [ProviderType.FACEBOOK]: { displayName: 'Facebook', path: '', icon: FacebookLogo },
 };
 
 const LoginPage = () => {
